RAG/requirements.txt
python-dotenv>=1.0.1
numpy>=1.26.0
scikit-learn>=1.4.0
pymongo>=4.6.0
sentence-transformers>=2.7.0
transformers>=4.40.0
torch>=2.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-google-genai>=1.0.0
faiss-cpu>=1.8.0
